requirements.txt
pandas==2.1.4  # use pandas <2.2.0 to avoid conflict with SQLAlchemy==1.4.51 when you connect to local mysql from docker conatainer
SQLAlchemy==1.4.51 # python3.12 latest supports 2.0.29, but airflow only supports up
jupyter==1.0.0
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5
bs4==0.0.2
apache-airflow==2.9.1
urllib3==1.26.6
openai-whisper==20231117
pytubefix==6.9.2
discord-webhook==1.3.1